Refactor send and receive modal
