+++ src/components/modals/wallet/index.tsx
@@ -1,0 +1,4 @@
+import ReceiveModal from "./ReceiveModal";
+import SendModal from "./SendModal";
+
+export { ReceiveModal, SendModal };
